renamed files + fixed a css bug
