+++ src/routes/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
-import Header from '../components/client/common/header/Header'
+import Header from '../components/client/common/header/'
 import LandingPage from '../pages/LandingPage'
 
 function RouteComponent() {
